+++ neosvr_headless_webui/dashboard.py
@@ -11,6 +11,14 @@ def dashboard():
     conn = connect(
         current_app.config["MANAGER_HOST"], current_app.config["MANAGER_PORT"]
     )
-    g.clients = conn.root.list_headless_clients()
+    clients = conn.root.list_headless_clients()
+
+    # Filter out clients that are not ready yet.
+    g.clients = {}
+    for c in clients:
+        if not clients[c].ready.is_set():
+            continue
+        g.clients[c] = clients[c]
+
     g.status = conn.root.get_manager_status()
     return render_template("dashboard.html")